Split cer and schema
