+++ app/reducers/followers.js
@@ -2,6 +2,10 @@
 
 import { createSelector } from 'reselect';
 import { selectCurrentUser } from './auth';
+import { Event } from 'app/actions/ActionTypes';
+import createEntityReducer, {
+  type EntityReducerTypes,
+} from 'app/utils/createEntityReducer';
 
 export const selectFollowersCurrentUser = createSelector(
   (state, props) =>
@@ -13,10 +17,7 @@ export const selectFollowersCurrentUser = createSelector(
 );
 
 export const selectFollowers = createSelector(
-  (state, { type }) =>
-    state[
-      'followers' + type.charAt(0).toUpperCase() + type.substr(1).toLowerCase()
-    ],
+  (state, { type }) => state[followersKeyGen(type)],
   (state, { target }) => target,
   (state, { follower }) => follower,
   (followers, target, follower) =>
@@ -28,3 +29,28 @@ export const selectFollowers = createSelector(
       )
     ]
 );
+
+export const followersKeyGen = (key: string) =>
+  'followers' + key.charAt(0).toUpperCase() + key.substr(1).toLowerCase();
+
+const followersSchemaGenerator = (
+  title,
+  types: {
+    fetch?: EntityReducerTypes,
+    mutate?: EntityReducerTypes,
+    delete?: EntityReducerTypes,
+  } = {}
+) =>
+  createEntityReducer({
+    key: followersKeyGen(title),
+    types,
+  });
+
+export const followersEvent = followersSchemaGenerator('Event', {
+  mutate: [Event.FOLLOW],
+  delete: [Event.UNFOLLOW],
+});
+
+export const followersCompany = followersSchemaGenerator('Company');
+
+export const followersUser = followersSchemaGenerator('User');
